Update read_lesion_imgs.m
In the previous version, the displayed filenames were wrong if excluded participants were more than 1.
Moreover the excluded images and scores were also wrong: e.g. If sub_idx = [3 7], then the loop excluded the first (not the third) subject (ni==1), and after the first was exluded, the loop excluded nii==2, which was the third subject before excluding the first.